+++ packages/core-ui/src/modules/layout/components/navigation/constants.ts
@@ -83,6 +83,12 @@ export const GENERAL_SETTINGS = [
     icon: 'icon-mail-alt',
     url: '/settings/email-templates',
   },
+  {
+    name: 'General Settings',
+    text: 'Exchange Rates',
+    icon: 'icon-mail-alt',
+    url: '/settings/exchangeRates',
+  },
 ];
 
 export const ACTIONS = [