+++ packages/core/src/data/permissions/actions/permission.ts
@@ -503,6 +503,26 @@ export const moduleObjects = {
     ],
   },
 
+  exchangeRates: {
+    name: 'exchangeRates',
+    description: 'ExchangeRates',
+    actions: [
+      {
+        name: 'exchangeRatesAll',
+        description: 'All',
+        use: ['showExchangeRates', 'manageExchangeRates'],
+      },
+      {
+        name: 'manageExchangeRates',
+        description: 'Manage exchange rates',
+      },
+      {
+        name: 'showExchangeRates',
+        description: 'Show exchange rates',
+      },
+    ],
+  },
+
   clients: {
     name: 'clients',
     description: 'Clients',