+++ packages/core/src/data/resolvers/queries/index.ts
@@ -33,6 +33,7 @@ import dashboards from './dashboard';
 import reports from './report';
 import sections from './section';
 import insights from './insight';
+import exchangeRates from './exchangeRates';
 
 export default {
   ...users,
@@ -66,4 +67,5 @@ export default {
   ...reports,
   ...sections,
   ...insights,
+  ...exchangeRates,
 };