+++ packages/plugin-inbox-ui/src/settings/integrations/components/messenger/Form.tsx
@@ -3,26 +3,27 @@ import { Appearance, Availability, Greeting, Intro, Options } from "./steps";
 import {
   Content,
   LeftContent,
-  MessengerPreview
+  MessengerPreview,
 } from "@erxes/ui-inbox/src/settings/integrations/styles";
 import {
   ControlWrapper,
   Indicator,
   Preview,
-  StepWrapper
+  StepWrapper,
 } from "@erxes/ui/src/components/step/styles";
 import {
+  IExternalLink,
   IIntegration,
   IMessages,
   IMessengerApps,
   IMessengerData,
   ISkillData,
   IUiOptions,
-  IExternalLink
 } from "@erxes/ui-inbox/src/settings/integrations/types";
 import { Step, Steps } from "@erxes/ui/src/components/step";
 
 import AddOns from "../../containers/messenger/AddOns";
+import BotSelector from "./steps/BotSelector";
 import Button from "@erxes/ui/src/components/Button";
 import CommonPreview from "./widgetPreview/CommonPreview";
 import Connection from "./steps/Connection";
@@ -52,11 +53,20 @@ type Props = {
   isLoading: boolean;
 };
 
+type BotPersistentMenuTypeMessenger = {
+  _id: string;
+  type: string;
+  text: string;
+  link: string;
+};
+
 type State = {
   title: string;
   botEndpointUrl?: string;
   botShowInitialMessage?: boolean;
   botCheck?: boolean;
+  botGreetMessage?: string;
+  persistentMenus?: BotPersistentMenuTypeMessenger[];
   skillData?: ISkillData;
   brandId: string;
   channelIds: string[];
@@ -109,7 +119,9 @@ class CreateMessenger extends React.Component<Props, State> {
       showVideoCallRequest: false,
       botEndpointUrl: "",
       botShowInitialMessage: false,
-      botCheck: false
+      botCheck: false,
+      botGreetMessage: "",
+      persistentMenus: [] as BotPersistentMenuTypeMessenger[],
     };
     const links = configData.links || {};
     const externalLinks = configData.externalLinks || [];
@@ -122,6 +134,8 @@ class CreateMessenger extends React.Component<Props, State> {
       title: integration.name,
       botEndpointUrl: configData.botEndpointUrl,
       botCheck: configData.botCheck,
+      botGreetMessage: configData.botGreetMessage,
+      persistentMenus: configData.persistentMenus,
       botShowInitialMessage: configData.botShowInitialMessage,
       skillData: configData.skillData,
       brandId: integration.brandId || "",
@@ -144,7 +158,7 @@ class CreateMessenger extends React.Component<Props, State> {
       showTimezone: configData.showTimezone || false,
       onlineHours: (configData.onlineHours || []).map((h) => ({
         _id: Math.random(),
-        ...h
+        ...h,
       })),
       showVideoCallRequest: configData.showVideoCallRequest,
       logo: uiOptions.logo || "",
@@ -156,7 +170,7 @@ class CreateMessenger extends React.Component<Props, State> {
       youtube: links.youtube || "",
       messages: { ...this.generateMessages(messages) },
       messengerApps,
-      externalLinks
+      externalLinks,
     };
   }
 
@@ -171,11 +185,11 @@ class CreateMessenger extends React.Component<Props, State> {
           title:
             message && message.greetings ? message.greetings.title || "" : "",
           message:
-            message && message.greetings ? message.greetings.message || "" : ""
+            message && message.greetings ? message.greetings.message || "" : "",
         },
         welcome: message.welcome || "",
         away: message.away || "",
-        thank: message.thank || ""
+        thank: message.thank || "",
       };
     });
 
@@ -188,6 +202,10 @@ class CreateMessenger extends React.Component<Props, State> {
 
   onExternalLinksChange = (newExternalLinks: IExternalLink[]) => {
     this.setState({ externalLinks: newExternalLinks });
+  };
+
+  onChangePersistent = (persistentMenus: BotPersistentMenuTypeMessenger[]) => {
+    this.setState({ persistentMenus: persistentMenus });
   };
 
   handleMessengerApps = (messengerApps: IMessengerApps) => {
@@ -203,6 +221,8 @@ class CreateMessenger extends React.Component<Props, State> {
       botShowInitialMessage,
       brandId,
       botCheck,
+      botGreetMessage,
+      persistentMenus,
       languageCode,
       channelIds,
       messages,
@@ -218,7 +238,7 @@ class CreateMessenger extends React.Component<Props, State> {
       showVideoCallRequest,
       messengerApps,
       skillData,
-      externalLinks
+      externalLinks,
     } = this.state;
 
     if (!languageCode) {
@@ -273,7 +293,7 @@ class CreateMessenger extends React.Component<Props, State> {
       facebook: linkify(facebook),
       instagram: linkify(instagram),
       twitter: linkify(twitter),
-      youtube: linkify(youtube)
+      youtube: linkify(youtube),
     };
 
     this.props.save({
@@ -286,6 +306,8 @@ class CreateMessenger extends React.Component<Props, State> {
         botEndpointUrl,
         botShowInitialMessage,
         botCheck,
+        botGreetMessage,
+        persistentMenus,
         notifyCustomer: this.state.notifyCustomer,
         availabilityMethod: this.state.availabilityMethod,
         isOnline: this.state.isOnline,
@@ -295,7 +317,7 @@ class CreateMessenger extends React.Component<Props, State> {
         onlineHours: (this.state.onlineHours || []).map((oh) => ({
           day: oh.day,
           from: oh.from,
-          to: oh.to
+          to: oh.to,
         })),
         supporterIds: this.state.supporterIds,
         messages,
@@ -306,15 +328,15 @@ class CreateMessenger extends React.Component<Props, State> {
         forceLogoutWhenResolve,
         showVideoCallRequest,
         links,
-        externalLinks
+        externalLinks,
       },
       uiOptions: {
         color: this.state.color,
         textColor: this.state.textColor,
         wallpaper: this.state.wallpaper,
-        logo: this.state.logo
+        logo: this.state.logo,
       },
-      messengerApps
+      messengerApps,
     });
   };
 
@@ -325,7 +347,7 @@ class CreateMessenger extends React.Component<Props, State> {
         name === "hours" ||
         name === "addon"
       ),
-      activeStep: name
+      activeStep: name,
     });
   };
 
@@ -333,10 +355,8 @@ class CreateMessenger extends React.Component<Props, State> {
     const { isLoading } = this.props;
 
     const cancelButton = (
-      <Link to='/settings/integrations'>
-        <Button
-          btnStyle='simple'
-          icon='times-circle'>
+      <Link to="/settings/integrations">
+        <Button btnStyle="simple" icon="times-circle">
           Cancel
         </Button>
       </Link>
@@ -347,9 +367,10 @@ class CreateMessenger extends React.Component<Props, State> {
         {cancelButton}
         <Button
           disabled={isLoading}
-          btnStyle='success'
+          btnStyle="success"
           icon={isLoading ? undefined : "check-circle"}
-          onClick={this.save}>
+          onClick={this.save}
+        >
           {isLoading && <SmallLoader />}
           Save
         </Button>
@@ -363,6 +384,8 @@ class CreateMessenger extends React.Component<Props, State> {
       botEndpointUrl,
       botShowInitialMessage,
       botCheck,
+      botGreetMessage,
+      persistentMenus,
       supporterIds,
       isOnline,
       availabilityMethod,
@@ -394,7 +417,7 @@ class CreateMessenger extends React.Component<Props, State> {
       channelIds,
       skillData,
       messengerApps,
-      externalLinks
+      externalLinks,
     } = this.state;
 
     const { integration } = this.props;
@@ -403,22 +426,20 @@ class CreateMessenger extends React.Component<Props, State> {
     const breadcrumb = [
       { title: __("Settings"), link: "/settings" },
       { title: __("Integrations"), link: "/settings/integrations" },
-      { title: __("Messenger") }
+      { title: __("Messenger") },
     ];
 
     return (
       <StepWrapper>
-        <Wrapper.Header
-          title={__("Messenger")}
-          breadcrumb={breadcrumb}
-        />
+        <Wrapper.Header title={__("Messenger")} breadcrumb={breadcrumb} />
         <Content>
           <LeftContent>
             <Steps>
               <Step
-                img='/images/icons/erxes-04.svg'
-                title='Appearance'
-                onClick={this.onStepClick.bind(null, "appearance")}>
+                img="/images/icons/erxes-04.svg"
+                title="Appearance"
+                onClick={this.onStepClick.bind(null, "appearance")}
+              >
                 <Appearance
                   onChange={this.onChange}
                   color={color}
@@ -429,9 +450,10 @@ class CreateMessenger extends React.Component<Props, State> {
               </Step>
 
               <Step
-                img='/images/icons/erxes-09.svg'
-                title='Greeting'
-                onClick={this.onStepClick.bind(null, "greeting")}>
+                img="/images/icons/erxes-09.svg"
+                title="Greeting"
+                onClick={this.onStepClick.bind(null, "greeting")}
+              >
                 <Greeting
                   teamMembers={this.props.teamMembers}
                   onChange={this.onChange}
@@ -448,9 +470,10 @@ class CreateMessenger extends React.Component<Props, State> {
               </Step>
 
               <Step
-                img='/images/icons/erxes-07.svg'
-                title='Intro'
-                onClick={this.onStepClick.bind(null, "intro")}>
+                img="/images/icons/erxes-07.svg"
+                title="Intro"
+                onClick={this.onStepClick.bind(null, "intro")}
+              >
                 <Intro
                   skillData={skillData}
                   onChange={this.onChange}
@@ -460,9 +483,10 @@ class CreateMessenger extends React.Component<Props, State> {
               </Step>
 
               <Step
-                img='/images/icons/erxes-03.svg'
+                img="/images/icons/erxes-03.svg"
                 title={__("Hours & Availability")}
-                onClick={this.onStepClick.bind(null, "hours")}>
+                onClick={this.onStepClick.bind(null, "hours")}
+              >
                 <Availability
                   onChange={this.onChange}
                   isOnline={isOnline}
@@ -476,9 +500,10 @@ class CreateMessenger extends React.Component<Props, State> {
               </Step>
 
               <Step
-                img='/images/icons/erxes-06.svg'
-                title='Default Settings'
-                onClick={this.onStepClick.bind(null, "default")}>
+                img="/images/icons/erxes-06.svg"
+                title="Default Settings"
+                onClick={this.onStepClick.bind(null, "default")}
+              >
                 <Options
                   onChange={this.onChange}
                   notifyCustomer={notifyCustomer}
@@ -490,11 +515,24 @@ class CreateMessenger extends React.Component<Props, State> {
                   showVideoCallRequest={showVideoCallRequest}
                 />
               </Step>
-
               <Step
-                img='/images/icons/erxes-16.svg'
+                img="/images/icons/erxes-24.svg"
+                title={__("Bot Setup")}
+                onClick={this.onStepClick.bind(null, "bot")}
+              >
+                <BotSelector
+                  title={title}
+                  botCheck={botCheck}
+                  botGreetMessage={botGreetMessage}
+                  persistentMenus={persistentMenus}
+                  onChange={this.onChange as any} // Explicitly cast
+                />
+              </Step>
+              <Step
+                img="/images/icons/erxes-16.svg"
                 title={__("Integration Setup")}
-                onClick={this.onStepClick.bind(null, "setup")}>
+                onClick={this.onStepClick.bind(null, "setup")}
+              >
                 <Connection
                   title={title}
                   botEndpointUrl={botEndpointUrl}
@@ -505,12 +543,12 @@ class CreateMessenger extends React.Component<Props, State> {
                   onChange={this.onChange}
                 />
               </Step>
-
               <Step
-                img='/images/icons/erxes-15.svg'
+                img="/images/icons/erxes-15.svg"
                 title={__("Add Ons")}
                 onClick={this.onStepClick.bind(null, "addon")}
-                noButton={true}>
+                noButton={true}
+              >
                 <AddOns
                   selectedBrand={brandId}
                   websiteMessengerApps={