+++ pos/app/reciept/layout.tsx
@@ -10,12 +10,12 @@ const PrintLayout = ({ children }: { children: React.ReactNode }) => {
 
   return (
     <div
-      className={cn(
-        "m-2 w-[72mm] space-y-1 p-1 pb-4 text-[10px] font-light shadow-lg print:m-0 print:pb-1 print:shadow-none border-b",
-        mode === "mobile" &&
-          "w-auto print:w-[72mm] font-normal print:font-light"
-      )}
-    >
+    className={cn(
+      "m-2 w-[72mm] relative overflow-y-auto min-h-screen space-y-1 p-1 pb-4 text-[10px] font-light shadow-lg print:m-0 print:pb-1 print:shadow-none print:h-auto print:overflow-visible border-b",
+      mode === "mobile" &&
+      "w-auto print:w-[72mm] font-normal print:font-light"
+  )}
+  >
       {children}
     </div>
   )